refactor NavBarMUI as separate component
